fix: PR 리뷰 지적 사항 수정
- GoogleComputeRoutesResponse: duration 소수 초 파싱 오류 수정 (Integer.parseInt → Double.parseDouble 후 int 캐스트)
- GoogleClientConfig: googleRoutesRestClient에 연결/읽기 타임아웃 추가 (5s/10s)
- RouteService: routes 빈 응답 방어 처리 및 @Cacheable 캐시 키 null 불일치 수정
- ScheduleItemController: /route 쿼리 파라미터 travelMode에 @Pattern 검증 추가
- ScheduleItem: updateTravelMode에 허용 값 검증 추가
- 테스트: 인가 verify, MEMBER 케이스, 신규 엔드포인트/서비스 메서드 테스트 추가

+++ src/main/java/com/howaboutus/backend/common/integration/google/dto/GoogleComputeRoutesResponse.java
@@ -13,7 +13,7 @@ public record GoogleComputeRoutesResponse(
         // Google returns duration as "Xs" string (e.g. "900s")
         public int durationSeconds() {
             if (duration == null || duration.isEmpty()) return 0;
-            return Integer.parseInt(duration.replace("s", ""));
+            return (int) Double.parseDouble(duration.replace("s", ""));
         }
     }
 }
